.claude/skills/strands-agents/assets/production/requirements.txt
strands-agents[bedrock]
strands-agents-tools
pydantic>=2.0
